Adds default image prop to HTMLHead
Adds a default value to the image prop in the HTMLHead component.

This ensures that a default image is always provided for social media meta tags, even when a specific image is not available for a page, preventing potential issues with social media sharing and SEO.

Sets a default image for Eleven Labs page.